sum of ints and floats

+++ 0x00-python_variable_annotations/6-sum_mixed_list.py
@@ -3,9 +3,9 @@
 from typing import List, Union
 
 
-def sum_mixed_list(mxd_list: List[Union[float, int]]) -> float:
+def sum_mixed_list(mxd_list: List[Union[int, float]]) -> float:
     """define func"""
-    j = 0.0
+    j = 0
     for i in mxd_list:
         j += i
     return j
